Use chrome storage API for icons
